requirements.txt
# Core deep learning
torch>=2.1
torchvision>=0.16

# NLP
transformers>=4.42
datasets>=2.19
sentencepiece>=0.1.99

# Utilities
numpy>=1.24
pillow>=10.0

# Optional but recommended
accelerate>=0.30
